move packet parser implementation from header to cpp files
