+++ include/l2/logical_link_control_packet.hpp
@@ -10,9 +10,7 @@
 
 #include "l2/upper_mac_packet.hpp"
 #include "utils/bit_vector.hpp"
-#include <bitset>
 #include <optional>
-#include <stdexcept>
 
 /// The type of the basic link packet.
 enum class BasicLinkType {
@@ -56,63 +54,10 @@ struct BasicLinkInformation {
     BasicLinkInformation() = delete;
 
     /// construct a BasicLinkInformation from a BitVector
-    explicit BasicLinkInformation(BitVector& data) {
-        auto pdu_type = data.take<4>();
-
-        switch (pdu_type) {
-        case 0b0000:
-            n_r_ = data.take<1>();
-            n_s_ = data.take<1>();
-            break;
-        case 0b0001:
-            n_s_ = data.take<1>();
-            break;
-        case 0b0010:
-            break;
-        case 0b0011:
-            n_r_ = data.take<1>();
-            break;
-        case 0b0100:
-            n_r_ = data.take<1>();
-            n_s_ = data.take<1>();
-            break;
-        case 0b0101:
-            n_s_ = data.take<1>();
-            break;
-        case 0b0110:
-            break;
-        case 0b0111:
-            n_r_ = data.take<1>();
-            break;
-        default:
-            throw std::runtime_error(
-                "Did not expect something other than basic link in LogicalLinkControlPacket parser!");
-        }
-
-        basic_link_type_ = BasicLinkType(pdu_type);
-
-        if (pdu_type >= 0b0100) {
-            auto fcs = data.take_last<32>();
-            auto computed_fcs = data.compute_fcs();
-            fcs_good_ = fcs == computed_fcs;
-        }
-    }
+    explicit BasicLinkInformation(BitVector& data);
 };
 
-inline auto operator<<(std::ostream& stream, const BasicLinkInformation& bli) -> std::ostream& {
-    stream << "  Basic Link Information: " << to_string(bli.basic_link_type_);
-    if (bli.n_r_) {
-        stream << " N(R): " << std::bitset<1>(*bli.n_r_);
-    }
-    if (bli.n_s_) {
-        stream << " N(S): " << std::bitset<1>(*bli.n_s_);
-    }
-    if (bli.fcs_good_) {
-        stream << " FCS good: " << (*bli.fcs_good_ ? "true" : "false");
-    }
-    stream << std::endl;
-    return stream;
-}
+auto operator<<(std::ostream& stream, const BasicLinkInformation& bli) -> std::ostream&;
 
 /// The packet that is parsed in the logical link control layer. Currently we only implement basic link.
 struct LogicalLinkControlPacket : public UpperMacCPlaneSignallingPacket {
@@ -121,23 +66,7 @@ struct LogicalLinkControlPacket : public UpperMacCPlaneSignallingPacket {
 
     LogicalLinkControlPacket() = delete;
 
-    explicit LogicalLinkControlPacket(const UpperMacCPlaneSignallingPacket& packet)
-        : UpperMacCPlaneSignallingPacket(packet) {
-        auto data = BitVector(*tm_sdu_);
-        auto pdu_type = data.look<4>(0);
-
-        /// We only implemented packet parsing for Basic Link PDUs at this point in time
-        if (pdu_type <= 0b1000) {
-            basic_link_information_ = BasicLinkInformation(data);
-            tl_sdu_ = data;
-        }
-    }
+    explicit LogicalLinkControlPacket(const UpperMacCPlaneSignallingPacket& packet);
 };
 
-inline auto operator<<(std::ostream& stream, const LogicalLinkControlPacket& llc) -> std::ostream& {
-    if (llc.basic_link_information_) {
-        stream << *llc.basic_link_information_;
-    }
-    stream << "  TL-SDU: " << llc.tl_sdu_ << std::endl;
-    return stream;
-}
+auto operator<<(std::ostream& stream, const LogicalLinkControlPacket& llc) -> std::ostream&;
